rename: type별로 폴더 생성 후 정리
