+++ src/pages/purchase/PurchaseOrderListPage.tsx
@@ -11,7 +11,7 @@ import type {
     PurchaseOrderSummary,
     PurchaseOrderDetail,
     PurchaseOrderStatus
-} from '@/types/purchase';
+} from '@/types/purchase/purchase.ts';
 import { Plus, FileText, FileDown, XCircle, Edit, X } from 'lucide-react';
 
 function StatusBadge({ status }: { status: PurchaseOrderStatus }) {
