+++ src/types/index.ts
@@ -1,19 +1,19 @@
-export * from './auth';
-export * from './user';
-export * from './store';
-export * from './invitation';
-export * from './ingredient';
-export * from './stock';
-export * from './common';
-export * from './product';
-export * from './vendor';
-export * from './menu';
-export * from './diningTable';
-export * from './sales';
-export * from './salesLedger';
-export * from './tableQr';
-export * from './purchase';
+export * from '@/types/user/auth.ts';
+export * from '@/types/user/user.ts';
+export * from './store/store.ts';
+export * from './store/invitation.ts';
+export * from './reference/ingredient.ts';
+export * from './stock/stock.ts';
+export * from './common/common.ts';
+export * from './stock/product.ts';
+export * from './reference/vendor.ts';
+export * from './reference/menu.ts';
+export * from './sales/diningTable.ts';
+export * from './sales/sales.ts';
+export * from './sales/salesLedger.ts';
+export * from './sales/tableQr.ts';
+export * from './purchase/purchase.ts';
 
-export * from "./ocr";
-export * from "./document";
-export * from "./disposal";
+export * from "./ocr/ocr.ts";
+export * from "./ocr/document.ts";
+export * from "./stock/disposal.ts";
