+++ src/types/ocr/document.ts
