+++ src/types/ocr/ocr.ts
