+++ src/types/sales/tableQr.ts
