+++ src/types/stock/product.ts
