+++ src/types/store/store.ts
